+++ lib/servo/include/Servo.h
@@ -11,9 +11,9 @@ class Servo {
   static constexpr int PWM_FREQ     = 50;   // DEFAULT PWM FREQUENCY, should work for both length and positional control
   static constexpr auto DEFAULT_MAX = 2ms;  // DEFAULT MAX WAVE LENGTH
   static constexpr auto DEFAULT_MIN = 1ms;  // DEFAULT MIN WAVE LENGTH
-  static constexpr std::chrono::duration<double> PERIOD = 1.0s / PWM_FREQ;  // DEFAULT PERIOD LENGTH
+  static constexpr std::chrono::duration<float> PERIOD = 1.0s / PWM_FREQ;  // DEFAULT PERIOD LENGTH
 
-  std::chrono::duration<double> m_max_pulse,  // PULSE LENGTH for MAX ANGLE/SPEED
+  std::chrono::duration<float> m_max_pulse,  // PULSE LENGTH for MAX ANGLE/SPEED
       m_min_pulse;                            // PULSE LENGTH for MIN ANGLE/SPEED
 
   int getSign(int val) {
@@ -25,35 +25,35 @@ class Servo {
   virtual ~Servo(){};
 
   // Optionally implemented
-  virtual mbed_error_status_t setRange(double range) {
+  virtual mbed_error_status_t setRange(float range) {
     return MBED_ERROR_INVALID_OPERATION;
   };
 
-  virtual mbed_error_status_t setMaxSpeed(double max_speed) {
+  virtual mbed_error_status_t setMaxSpeed(float max_speed) {
     return MBED_ERROR_INVALID_OPERATION;
   };
 
-  virtual double getRange(void) {
+  virtual float getRange(void) {
     return -1;
   };
 
-  virtual double getMaxSpeed(void) {
+  virtual float getMaxSpeed(void) {
     return -1;
   };
 
-  virtual mbed_error_status_t setPosition(double angle) {
+  virtual mbed_error_status_t setPosition(float angle) {
     return MBED_ERROR_INVALID_OPERATION;
   };
 
-  virtual mbed_error_status_t setSpeed(double speed) {
+  virtual mbed_error_status_t setSpeed(float speed) {
     return MBED_ERROR_INVALID_OPERATION;
   };
 
   // Must be implemented
-  virtual double read(void) = 0;
+  virtual float read(void) = 0;
 
   // Override default period (ONLY USE FOR SPECIFIC FREQ REQUIREMENT)
-  void setPeriod(std::chrono::duration<double> period) {
+  void setPeriod(std::chrono::duration<float> period) {
     m_pwm.period(period.count());
   };
 };
